upd: lint struct size
